+++ test_unstructured/staging/test_base_staging.py
@@ -128,5 +128,9 @@ def test_serialized_deserialize_elements_to_json(tmpdir):
     ]
 
     base.elements_to_json(elements, filename=filename)
-    new_elements = base.elements_from_json(filename=filename)
-    assert elements == new_elements
+    new_elements_filename = base.elements_from_json(filename=filename)
+    assert elements == new_elements_filename
+
+    elements_str = base.elements_to_json(elements)
+    new_elements_text = base.elements_from_json(text=elements_str)
+    assert elements == new_elements_text